fix: sync correct extensions/pwpush files and remove root-level duplicates
- Update extensions/pwpush/ with latest local changes
- Remove stray src/ files created by earlier sync
- Address Greptile feedback: server resolution, payload whitespace, history keys, per-record API keys
